refactor: Tidy up ts files into proxy dir
