+++ src/adaptor/index.ts
@@ -3,10 +3,8 @@ export * from './discord/voice.js';
 export * from './embed-convert.js';
 export * from './in-memory-reservation.js';
 export * from './in-memory-typo.js';
-export * from './message-proxy.js';
 export * from './mock-voice.js';
+export * from './proxy.js';
 export * from './random.js';
-export * from './transformer.js';
-export * from './voice-room-proxy.js';
 export * from './discord/participant.js';
 export * from './discord/output.js';
