+++ src/adaptor/proxy.ts
@@ -1,0 +1,5 @@
+export * from './proxy/emoji.js';
+export * from './proxy/message.js';
+export * from './proxy/role.js';
+export * from './proxy/transformer.js';
+export * from './proxy/voice-room.js';
