+++ src/adaptor/proxy/transformer/bot-filter.ts
