+++ src/adaptor/proxy/transformer/tuple.ts
@@ -10,7 +10,7 @@
         \- A2=> <----------- B2=> <-/
 */
 
-import type { MessageHandler } from '../index.js';
+import type { MessageHandler } from '../../index.js';
 import type { Transformer } from '../transformer.js';
 
 // A1=> -> A2=> -> A1*A2=>
